217、233、400、887

+++ jianzhioffer/src/com/zhang/java/Offer44.java
@@ -3,7 +3,7 @@ package com.zhang.java;
 /**
  * @Date 2022/3/27 8:49
  * @Author zsy
- * @Description 数字序列中某一位的数字 类比Offer43
+ * @Description 数字序列中某一位的数字 类比Offer43、Problem233 同Problem400
  * 数字以0123456789101112131415…的格式序列化到一个字符序列中。
  * 在这个序列中，第5位（从下标0开始计数）是5，第13位是1，第19位是4，等等。
  * 请写一个函数，求任意第n位对应的数字。
@@ -24,16 +24,16 @@ public class Offer44 {
 
     /**
      * 模拟
-     * 每次判断长度为length的数字是否在n范围之内，不在则往后继续寻找，找到则在长度为length的数字中查找
-     * 找到n是属于k位数字的范围，从k位数字的起始位置找到n所属的数字，找到对应下标
+     * 计算长度为1、2、3...数字的个数，找到n所在长度为几的数字中，确定所在位置
+     * 时间复杂度O(logn)，空间复杂度O(logn)
+     * <p>
      * 例如：n=123
      * 1位数字范围共9个数字，不属于1位数字
      * 2位数字范围共9+180个数字，属于2位数字范围
      * 123-9-1=113
      * 113/2=56
      * 113%2=1
-     * 说明，123是2位数字66的第2位，即为6
-     * 时间复杂度O(logn)，空间复杂度O(logn)
+     * 说明，123是2位数字10+56=66的下标索引为1的数字，即为6
      * <p>
      * 范围         个数        所占的位数
      * 1-9          9           9*1=9
@@ -54,11 +54,11 @@ public class Offer44 {
         //当前数字的位数
         int length = 1;
         //长度为length数字的个数
-        int size = 9;
+        long size = 9;
 
         //使用long，避免相乘溢出
-        while (count + (long) size * length < n) {
-            count = count + (long) size * length;
+        while (count + size * length < n) {
+            count = count + size * length;
             length++;
             size = size * 10;
         }
